Refactor Tests
- update namespace from `Test` to `Tests`
- namespace models into `Tests\TestModels`
- update PHPUnit setUp() fixture (protected, not public)

Thanks to @AyoobMH  - ref PR #2240
