+++ composer.json
@@ -45,7 +45,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Spatie\\Permission\\Test\\": "tests"
+            "Spatie\\Permission\\Tests\\": "tests"
         }
     },
     "config": {
