+++ tests/BladeTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Spatie\Permission\Test;
+namespace Spatie\Permission\Tests;
 
-use Artisan;
+use Illuminate\Support\Facades\Artisan;
 use Spatie\Permission\Contracts\Role;
 
 class BladeTest extends TestCase
 {
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
