+++ tests/CommandTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Spatie\Permission\Test;
+namespace Spatie\Permission\Tests;
 
 use Illuminate\Support\Facades\Artisan;
 use Spatie\Permission\Models\Permission;
