+++ tests/RouteTest.php
@@ -1,8 +1,6 @@
 <?php
 
-namespace Spatie\Permission\Test;
-
-use Illuminate\Http\Response;
+namespace Spatie\Permission\Tests;
 
 class RouteTest extends TestCase
 {
@@ -47,22 +45,5 @@ class RouteTest extends TestCase
             ],
             $this->getLastRouteMiddlewareFromRouter($router)
         );
-    }
-
-    protected function getLastRouteMiddlewareFromRouter($router)
-    {
-        return last($router->getRoutes()->get())->middleware();
-    }
-
-    protected function getRouter()
-    {
-        return app('router');
-    }
-
-    protected function getRouteResponse()
-    {
-        return function () {
-            return (new Response())->setContent('<html></html>');
-        };
     }
 }
