+++ tests/TestModels/Role.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Spatie\Permission\Test;
+namespace Spatie\Permission\Tests\TestModels;
 
 use Illuminate\Database\Eloquent\SoftDeletes;
