+++ tests/TestModels/RuntimeRole.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Spatie\Permission\Test;
+namespace Spatie\Permission\Tests\TestModels;
 
 class RuntimeRole extends \Spatie\Permission\Models\Role
 {
