+++ tests/TestModels/User.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Spatie\Permission\Test;
+namespace Spatie\Permission\Tests\TestModels;
 
 use Illuminate\Auth\Authenticatable;
 use Illuminate\Contracts\Auth\Access\Authorizable as AuthorizableContract;
@@ -9,7 +9,7 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Foundation\Auth\Access\Authorizable;
 use Spatie\Permission\Traits\HasRoles;
 
-class Manager extends Model implements AuthorizableContract, AuthenticatableContract
+class User extends Model implements AuthorizableContract, AuthenticatableContract
 {
     use HasRoles;
     use Authorizable;
@@ -20,14 +20,4 @@ class Manager extends Model implements AuthorizableContract, AuthenticatableCont
     public $timestamps = false;
 
     protected $table = 'users';
-
-    // this function is added here to support the unit tests verifying it works
-    // When present, it takes precedence over the $guard_name property.
-    public function guardName()
-    {
-        return 'jwt';
-    }
-
-    // intentionally different property value for the sake of unit tests
-    protected $guard_name = 'web';
 }
