+++ tests/WildcardRoleTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Spatie\Permission\Test;
+namespace Spatie\Permission\Tests;
 
 use Spatie\Permission\Models\Permission;
 
 class WildcardRoleTest extends TestCase
 {
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
